requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.21
redis==5.0.1